added pages for each service, as well as links between Advisors and clients involved

+++ app/models/ClientForm.scala
@@ -2,7 +2,7 @@ package models
 
 import java.util.Date
 
-case class ClientForm(cid: String, name: String, birthdate: Date, homephone: Int, workphone: Int,
+case class ClientForm(cid: String, name: String, birthdate: Date, homephone: Long, workphone: Long,
                       streetAddress: String, city: String, province: String, postalCode: String) {
 
 }
